fuzzy_search/requirements.txt
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.2
django-cors-headers==2.4.0
djangorestframework==3.9.0
fuzzywuzzy==0.17.0
idna==2.7
python-Levenshtein==0.12.0
